pyproject.toml
[tool.poetry]
name = "songdatabase"
version = "0.1.0"
description = "CS 3310 Project: Creating a relational database application to manage a song collection."
authors = ["Brent Bradford <brent.c.bradford@gmail.com>"]
license = "apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiosqlite = "^0.20.0"
pyqt6 = "^6.7.1"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[tool.poetry.scripts]
songdatabase = "songdatabase.app:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
